requirements.txt
asgiref==3.3.1
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.4
django-cleanup==5.1.0
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
